==> manuscript/Annotation and loading/Step_4_combine_and_clean-github.ipynb <==
Merge object posty cleanup


# fine annotate lineage 
import scanpy as sc
import numpy as np
import pandas as pd
import scvi
import seaborn as sns
import os,sys
#import mudata
import anndata
## Add this line so the text on pdf is correctly recognised!!!
from matplotlib import rcParams
rcParams['pdf.fonttype'] = 42
import scvi
import scvi_wrapper as sv
import anndata as ad

#import multi_view_atlas as mva
sc.logging.print_header()
sc.settings.set_figure_params(dpi = 150, color_map = 'RdPu', dpi_save = 300, vector_friendly = True, format = 'pdf')

# Load large object pre final cleanup 

adata = sc.read_h5ad('...adata_full_rev_2.h5ad')
adata

sc.set_figure_params(fontsize=10,figsize=[10,10],dpi=100)
sc.pl.umap(adata,
           color=['leiden_2','cell_type_level_0'], 
           frameon=False,
           s=1,
           # groups=['10'],
           legend_loc='on data',
           legend_fontoutline=1,
           outline_color='white',
          )

# combine all cleaned object 

adata_dn = sc.read_h5ad('...clean/adata_dn_rev_1_clean.h5ad')
adata_dn

adata_dp = sc.read_h5ad('...clean/adata_dp_rev_1_clean.h5ad')
adata_dp

adata_mye = sc.read_h5ad('...clean/adata_mye_rev_1_clean.h5ad')
adata_mye

adata_str = sc.read_h5ad('...clean/adata_str_rev_1_clean.h5ad')
adata_str

adata_sp = sc.read_h5ad('...clean/adata_t_sp_rev_1_clean.h5ad')
adata_sp

adata_tec = sc.read_h5ad('...clean/adata_tec_rev_1_clean.h5ad')
adata_tec

adata_vas = sc.read_h5ad('...clean/adata_vas_rev_1_clean.h5ad')
adata_vas

adata_new = ad.concat([adata_dn,adata_dp,adata_mye,adata_str,adata_sp,adata_tec,adata_vas],join='outer')
adata_new

adata_new = sv.scvi_wrapper_basic(
    adata_new,
    max_epochs=350,
    remove_vdjgenes=True,
    hvg=5000,
    batch_key='sample',
    categorical_covariate_keys=['chemistry_simple','age_group','study','sex','donor'],
    continuous_covariate_keys = None,
    cluster_leiden=False,
    layer_use=None,
    remove_cite=False,
) 
sc.tl.leiden(adata_new, resolution=1, key_added="leiden_1")
sc.tl.leiden(adata_new, resolution=2, key_added="leiden_2")
sc.tl.leiden(adata_new, resolution=3, key_added="leiden_3")
sc.tl.leiden(adata_new, resolution=4, key_added="leiden_4")

sc.set_figure_params(fontsize=10,figsize=[10,10],dpi=100)
sc.pl.umap(adata_new,
           color=['leiden_2','donor','doublet_score','study','leiden_3','leiden_4'],
           frameon=False,
           ncols=2,
           s=5,
           legend_loc='on data',
           legend_fontoutline=1,
           outline_color='white',
          )

os.chdir('...clean/')
adata_new.write_h5ad('adata_full_rev_2_clean.h5ad')

adata_new.obs.to_csv('adata_full_rev_2_clean.csv')
